conversion into package mostly complete
still some errors with appearance dialog not accepting male/futa
renamed some files
updated gitignore
