+++ .gitignore
@@ -1,1 +1,5 @@
 /__pycache__/*.pyc
+/ecc/__pycache__/*.pyc
+/ecc/common/__pycache__/*.pyc
+/ecc/gui/__pycache__/*.pyc
+/ecc/logic/__pycache__/*.pyc
